Améliore le chargement des résultats transformés de la simulation
résultats transformés
